+++ src/lib/st-search/st-search.component.spec.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { DebugElement, SimpleChanges } from '@angular/core';
+import { DebugElement, SimpleChanges, SimpleChange } from '@angular/core';
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
@@ -21,6 +21,9 @@ import { By } from '@angular/platform-browser';
 // Component
 import { StSearchComponent } from './st-search.component';
 
+// Modules
+import { StDropdownMenuModule } from '../st-dropdown-menu/st-dropdown-menu.module';
+import { StDropDownMenuItem } from '../st-dropdown-menu/st-dropdown-menu.interface';
 
 let comp: StSearchComponent;
 let fixture: ComponentFixture<StSearchComponent>;
@@ -28,14 +31,12 @@ let de: DebugElement;
 
 let id: string = 'search-123';
 let placeholder: string = 'search a text';
-let debounce: number = 10;
-let minLength: number = 0;
 
 describe('StSearchComponent', () => {
 
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [FormsModule, ReactiveFormsModule],
+         imports: [FormsModule, ReactiveFormsModule, StDropdownMenuModule],
          declarations: [StSearchComponent]
       })
          .compileComponents();  // compile template and css
@@ -47,8 +48,6 @@ describe('StSearchComponent', () => {
 
       comp.qaTag = id;
       comp.placeholder = placeholder;
-      comp.debounce = debounce;
-      comp.minLength = minLength;
    });
 
    it('should be initialized correctly', () => {
@@ -70,6 +69,7 @@ describe('StSearchComponent', () => {
    });
 
    it('should be search (async)', fakeAsync(() => {
+      comp.debounce = 10;
       fixture.detectChanges();
       let result: string = 'test';
       let input: HTMLInputElement = fixture.debugElement.query(By.css('input')).nativeElement;
@@ -103,8 +103,42 @@ describe('StSearchComponent', () => {
       expect(outputSearch).toEqual(result);
    }));
 
+   it('should be searched with new delay', fakeAsync(() => {
+      comp.debounce = 1000;
+      fixture.detectChanges();
+
+      let result: string = 'test';
+      let input: HTMLInputElement = fixture.debugElement.query(By.css('input')).nativeElement;
+
+      let outputSearch: string = '';
+      comp.search.subscribe((search: string) => outputSearch = search);
+      input.value = result;
+      input.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      tick(500);
+      expect(outputSearch).not.toEqual(result);
+      tick(501);
+      expect(outputSearch).toEqual(result);
+
+      let changes: SimpleChanges = { debounce: new SimpleChange(1000, 250, true) };
+      comp.debounce = 250;
+      comp.ngOnChanges(changes);
+      fixture.detectChanges();
+
+      outputSearch = '';
+      result = 'test2';
+      input.value = result;
+      input.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      tick(200);
+      expect(outputSearch).not.toEqual(result);
+      tick(55);
+      expect(outputSearch).toEqual(result);
+   }));
+
    it('should be searched with min length', fakeAsync(() => {
-      comp.debounce = 0;
       comp.minLength = 3;
       let responseFunction = jasmine.createSpy('response');
       comp.search.subscribe(responseFunction);
@@ -115,15 +149,15 @@ describe('StSearchComponent', () => {
       input.value = 'te';
       input.dispatchEvent(new Event('input'));
       fixture.detectChanges();
-      tick(10);
 
+      tick(1);
       expect(responseFunction).not.toHaveBeenCalled();
 
       input.value = 'test';
       input.dispatchEvent(new Event('input'));
       fixture.detectChanges();
-      tick(10);
 
+      tick(1);
       expect(responseFunction).toHaveBeenCalled();
       expect(responseFunction).toHaveBeenCalledWith('test');
    }));
@@ -162,96 +196,74 @@ describe('StSearchComponent', () => {
       expect(responseFunction).toHaveBeenCalled();
       expect(responseFunction).toHaveBeenCalledTimes(2);
       expect(responseFunction).toHaveBeenCalledWith('test');
+
+      input.triggerEventHandler('keypress', { which: 13 });
+      fixture.detectChanges();
+
+      expect(responseFunction).toHaveBeenCalled();
+      expect(responseFunction).toHaveBeenCalledTimes(3);
+      expect(responseFunction).toHaveBeenCalledWith('test');
    });
 
-   it('should be searched when click', () => {
-      comp.debounce = 0;
-      comp.minLength = 0;
+   it('should be searched twice', fakeAsync(() => {
       let responseFunction = jasmine.createSpy('response');
       comp.search.subscribe(responseFunction);
 
       fixture.detectChanges();
       let input: DebugElement = fixture.debugElement.query(By.css('input'));
-      let searchButton: DebugElement = fixture.debugElement.query(By.css('.st-search-icon'));
 
       input.nativeElement.value = 'te';
       input.nativeElement.dispatchEvent(new Event('input'));
-      searchButton.triggerEventHandler('mousedown', {});
       fixture.detectChanges();
 
+      tick(1);
       expect(responseFunction).toHaveBeenCalled();
       expect(responseFunction).toHaveBeenCalledTimes(1);
       expect(responseFunction).toHaveBeenCalledWith('te');
 
-      input.nativeElement.value = 'test';
+      input.nativeElement.value = 'te';
       input.nativeElement.dispatchEvent(new Event('input'));
-      searchButton.triggerEventHandler('mousedown', {});
       fixture.detectChanges();
 
+      tick(1);
+      expect(responseFunction).toHaveBeenCalledTimes(1);
+
+      input.nativeElement.value = 'test';
+      input.nativeElement.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      tick(1);
       expect(responseFunction).toHaveBeenCalled();
       expect(responseFunction).toHaveBeenCalledTimes(2);
       expect(responseFunction).toHaveBeenCalledWith('test');
-   });
+   }));
 
-   it('should be searched only on click', () => {
-      comp.debounce = 0;
-      comp.minLength = 0;
-      comp.searchOnlyOnClick = true;
+   it('should be searched when press enter without liveSearch', () => {
+      comp.liveSearch = false;
       let responseFunction = jasmine.createSpy('response');
       comp.search.subscribe(responseFunction);
 
       fixture.detectChanges();
       let input: DebugElement = fixture.debugElement.query(By.css('input'));
-      let searchButton: DebugElement = fixture.debugElement.query(By.css('.st-search-icon'));
+
+      input.nativeElement.value = 'te';
+      input.nativeElement.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      expect(responseFunction).not.toHaveBeenCalled();
 
       input.nativeElement.value = 'test';
       input.nativeElement.dispatchEvent(new Event('input'));
       input.triggerEventHandler('keypress', { which: 13 });
       fixture.detectChanges();
 
-      expect(responseFunction).not.toHaveBeenCalled();
-      expect(responseFunction).toHaveBeenCalledTimes(0);
-
-      searchButton.triggerEventHandler('mousedown', {});
-      fixture.detectChanges();
-
       expect(responseFunction).toHaveBeenCalled();
-      expect(responseFunction).toHaveBeenCalledTimes(1);
-      expect(responseFunction).toHaveBeenCalledWith('test');
-   });
-
-   it('should be searched on click twice the same', () => {
-      comp.debounce = 0;
-      comp.minLength = 0;
-      let responseFunction = jasmine.createSpy('response');
-      comp.search.subscribe(responseFunction);
-
-      fixture.detectChanges();
-      let input: DebugElement = fixture.debugElement.query(By.css('input'));
-      let searchButton: DebugElement = fixture.debugElement.query(By.css('.st-search-icon'));
-
-      input.nativeElement.value = 'test';
-      input.nativeElement.dispatchEvent(new Event('input'));
-
-      searchButton.triggerEventHandler('mousedown', {});
-      fixture.detectChanges();
-
-      expect(responseFunction).toHaveBeenCalled();
-      expect(responseFunction).toHaveBeenCalledTimes(1);
-      expect(responseFunction).toHaveBeenCalledWith('test');
-
-      searchButton.triggerEventHandler('mousedown', {});
-      fixture.detectChanges();
-
-      expect(responseFunction).toHaveBeenCalled();
-      expect(responseFunction).toHaveBeenCalledTimes(2);
       expect(responseFunction).toHaveBeenCalledWith('test');
    });
 
    it('should be able to clear on click', () => {
       comp.debounce = 0;
       comp.minLength = 0;
-      comp.hasClearButton = true;
 
       fixture.detectChanges();
 
@@ -269,27 +281,6 @@ describe('StSearchComponent', () => {
       expect(comp.searchBox.value).toEqual('');
    });
 
-
-   it('should be able to change focus', () => {
-      comp.debounce = 0;
-      comp.minLength = 0;
-      fixture.detectChanges();
-      let input: DebugElement = fixture.debugElement.query(By.css('input'));
-
-      expect(comp.focus).toBeFalsy();
-
-      input.nativeElement.dispatchEvent(new Event('focus'));
-      fixture.detectChanges();
-
-      expect(comp.focus).toBeTruthy();
-
-      input.nativeElement.dispatchEvent(new Event('blur'));
-      fixture.detectChanges();
-
-      expect(comp.focus).toBeFalsy();
-   });
-
-
    it('should be able to change initial Values', () => {
       comp.debounce = 0;
       comp.minLength = 0;
@@ -300,8 +291,8 @@ describe('StSearchComponent', () => {
       expect((<HTMLInputElement>input.nativeElement).value).toEqual('Initial value');
 
       let changes: SimpleChanges = {
-         value: { currentValue: 'new value', previousValue: 'Initial value', firstChange: true, isFirstChange: () => true },
-         liveSearch: { currentValue: false, previousValue: true, firstChange: true, isFirstChange: () => true }
+         value: new SimpleChange('Initial value', 'new value', true),
+         liveSearch: new SimpleChange(true, false, true)
       };
 
       comp.value = 'new value';
@@ -330,4 +321,107 @@ describe('StSearchComponent', () => {
       let disabledState: string = input.nativeElement.getAttribute('disabled');
       expect(disabledState).not.toBeNull();
    });
+
+   it('should be able to change to disabled state twice', () => {
+      comp.disabled = true;
+      fixture.detectChanges();
+      let input: DebugElement = fixture.debugElement.query(By.css('input'));
+
+      let disabledState: string = input.nativeElement.getAttribute('disabled');
+      expect(disabledState).not.toBeNull();
+
+      let changes: SimpleChanges = { disabled: new SimpleChange(true, false, true) };
+      comp.ngOnChanges(changes);
+      disabledState = input.nativeElement.getAttribute('disabled');
+      expect(disabledState).toEqual('');
+   });
+
+   it('should destroy correctly', () => {
+      comp.isActive = true;
+      expect(comp.isActive).toBeTruthy();
+      comp.ngOnDestroy();
+      expect(comp.isActive).toBeFalsy();
+   });
+
+   it('should be search with autocomplete', () => {
+
+      comp.autocompleteList = [{ value: '1', label: '1' }, { value: '2', label: '2' }];
+      comp.withAutocomplete = true;
+      let responseFunction = jasmine.createSpy('response');
+      comp.search.subscribe(responseFunction);
+
+      fixture.detectChanges();
+      expect(responseFunction).not.toHaveBeenCalled();
+
+      comp.changeOption(undefined);
+      expect(responseFunction).not.toHaveBeenCalled();
+
+      comp.changeOption(comp.autocompleteList[0]);
+      expect(responseFunction).toHaveBeenCalled();
+      expect(responseFunction).toHaveBeenCalledTimes(1);
+      expect(responseFunction).toHaveBeenCalledWith(comp.autocompleteList[0].label);
+   });
+
+   it('should be search with autocomplete', fakeAsync(() => {
+      comp.autocompleteList = [{ value: '1', label: '1' }, { value: '2', label: '2' }];
+      comp.withAutocomplete = true;
+      fixture.detectChanges();
+
+      let input: HTMLInputElement = fixture.debugElement.query(By.css('input')).nativeElement;
+      let dropdownItems: DebugElement[] = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+
+      expect(input).toBeDefined();
+      expect(dropdownItems.length).toEqual(0);
+
+
+      input.value = 'te';
+      input.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      tick(1);
+      fixture.detectChanges();
+      dropdownItems = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+      expect(dropdownItems.length).toEqual(2);
+
+
+      input.value = '';
+      input.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+      tick(1);
+      fixture.detectChanges();
+      dropdownItems = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+      expect(dropdownItems.length).toEqual(0);
+   }));
+
+   it('should be change autocomplete list', fakeAsync(() => {
+      let initialList: StDropDownMenuItem[] = [{ value: '1', label: '1' }, { value: '2', label: '2' }];
+      let finalList: StDropDownMenuItem[] = [{ value: '1', label: '1' }, { value: '2', label: '2' }, { value: '3', label: '3' }];
+      comp.autocompleteList = initialList;
+      comp.withAutocomplete = true;
+      fixture.detectChanges();
+
+      let input: HTMLInputElement = fixture.debugElement.query(By.css('input')).nativeElement;
+      let dropdownItems: DebugElement[] = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+
+      expect(input).toBeDefined();
+      expect(dropdownItems.length).toEqual(0);
+
+
+      input.value = 'te';
+      input.dispatchEvent(new Event('input'));
+      fixture.detectChanges();
+
+      tick(1);
+      fixture.detectChanges();
+      dropdownItems = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+      expect(dropdownItems.length).toEqual(initialList.length);
+
+
+      comp.autocompleteList = finalList;
+      let changes: SimpleChanges = { autocompleteList: new SimpleChange(initialList, finalList, true) };
+      comp.ngOnChanges(changes);
+      fixture.detectChanges();
+      dropdownItems = fixture.debugElement.queryAll(By.css('st-dropdown-menu-item'));
+      expect(dropdownItems.length).toEqual(finalList.length);
+   }));
 });